Improving results of refactoring
